Atualizacao dos testes de BookChapter e Conferencia
