+++ grails-app/i18n/messages_pt_BR.properties
@@ -1,4 +1,4 @@
-
+﻿
 default.doesnt.match.message=O campo [{0}] da classe [{1}] com o valor [{2}] não atende ao padrão definido [{3}]
 default.invalid.url.message=O campo [{0}] da classe [{1}] com o valor [{2}] não é uma URL válida
 default.invalid.creditCard.message=O campo [{0}] da classe [{1}] com o valor [{2}] não é um número válido de cartão de crédito
@@ -62,7 +62,7 @@ default.member.label=Membros
 default.publication.label=Publicações
 default.MSGIni.label=Escolha uma opção do menu acima
 default.periodico.label=Periódico
-default.conferencia.label=Conferência
+default.conferencia.label=Conferencia
 default.ferramenta.label=Ferramenta
 #if($maisresul)
 default.dissertacao.label=Dissertação
@@ -79,13 +79,13 @@ default.search.label=Busca
 
 conferencia.title.label=Título
 conferencia.author.label=Autor
-conferencia.conference.label=Conferência
+conferencia.conference.label=Conferencia
 conferencia.year.label=Ano
 conferencia.pageInitial.label=Página Inicial
 conferencia.pageFinal.label=Página Final
 conferencia.month.label=Mês
 conferencia.arquivo.label=Arquivo
-conferencia.label=Conferência
+conferencia.label=Conferencia
 conferencia.addAutor.label='Adicionar Autor'
 conferencia.members.label=Membros Autores
