+++ grails-app/views/bookChapter/_form.gsp
@@ -21,9 +21,8 @@
 <div class="fieldcontain ${hasErrors(bean: bookChapterInstance, field: 'file', 'error')} ">
 	<label for="file">
 		<g:message code="bookChapter.file.label" default="File" />
-		
 	</label>
-	<g:textArea name="file" cols="40" rows="5" maxlength="100000" value="${bookChapterInstance?.file}"/>
+	<g:field type="file" name="file" id="file" required="" value="${fieldValue(bean: bookChapterInstance, field: 'file')}"/>
 </div>
 
 <div class="fieldcontain ${hasErrors(bean: bookChapterInstance, field: 'researchLine', 'error')} ">
